requirements.txt
# Flask==2.0.1
# flask-cors==3.0.10
# flask-jwt-extended==4.4.4
# mysql-connector-python==8.0.31
# google-generativeai==0.3.1
# python-dotenv==0.21.0
# Werkzeug==2.0.2
# bcrypt==4.0.1
Flask
flask-cors
flask-jwt-extended
mysql-connector-python
google-generativeai
python-dotenv
Werkzeug
bcrypt
pandas
faiss-cpu
google-generativeai
langchain
langchain-core
# langchain-google-genai>=0.0.5
langchain-google-genai
langchain-community
langchain-qdrant
pydantic
openpyxl
gunicorn
PyMuPDF
qdrant-client
# grpcio-tools==1.59.0 --only-binary :all:
grpcio-tools --only-binary :all: